requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-ckeditor==6.0.0
django-crispy-forms==1.11.1
django-js-asset==1.2.2
django-request==1.5.6
django-seo==0.3.5
django-tawkto==0.3
django-tinymce==3.2.0
idna==2.10
Pillow==8.1.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
urllib3==1.26.3
